// web/src/api/types.gen.ts
export type PatchApiProductsByIdData = {
    body: {
        name?: string;
        sku?: string;
        category?: 'electronics' | 'clothing' | 'food' | 'furniture' | 'tools' | 'other';
        price?: number;
        quantity?: number;
        minStock?: number;
        storeId?: number;
    };
    path: {
        id: number;
    };
    query?: never;
    url: '/api/products/{id}';
};